Move RegisterBus impl in RegisterBus
